app: implement edit-contents functionality
Steal the (brilliant) <Ctrl><Shift>S shortcut from simple-term.  Our
implementation of it uses a pipe instead of a temporary file.

In testing, even with large buffers, that works great, but there's a
theoretical problem lurking here: if vim decided to want to interact
with the terminal emulator before reading from stdin, the pipe could
backup, blocking the terminal from responding.  If that happens, we can
switch to a memfd or an anonymous tempfile.