Major restructure of application model
Boxi now defaults to opening a terminal session on the host, using the
same agent model as when working with containers.  This works from
inside Flatpak as well.

If you want to run a command inside of a container you need to give the
`--container`/`-c` commandline argument.

The addition of local commandline handling introduced non-trivial
duplication between the "app" and "client" code, and bugs[1] in
pygobject prevent accomplishing this in a reasonable way with a mixin.
Drop the client code and add a new big of magic to the primary
Application object: we first try to register as launcher, falling back
to the primary instance if it fails.  With this hack in place, the main
advantage of having the separate client code has disappeared.  There is
still a small performance and memory hit to having to initialise Gtk
when using the launcher, however.

[1] https://gitlab.gnome.org/GNOME/pygobject/-/issues/512