agent: run as a daemon
The agent runs until the control socket (fd 3) is closed.  In theory
that happens when Boxi shuts down, but we send that socket through quite
a few different layers before it gets to the agent, and none of them
stop running for as long as the agent is running, creating a catch-22.

Let's have the agent daemonize itself so that `podman` (and other
intermediaries) can stop running, dropping all other references on fd 3.