agent: accept json object with "args" and "env"
Instead of sending the args as an array, send a dictionary with the args
and the environment variables to replace.

Use this to set the correct TERM and VTE_VERSION variables.  The later
is required to ensure we get VTE shell integration enabled.

At some point it will make sense to expand this to allow the user to
specify environment variable overrides.