app: set prgname to our appid
This results in our wmclass also being set to the application ID, which
helps gnome-shell find the desktop files installed for running Boxi with
a specific container.

This is needed for when we're running under Flatpak, where otherwise the
sandbox ID would take precedence, leading us to be identified as the
main app.

See https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/84