monitor: Automatically create desktop files
Add a new "monitor" command that uses `podman events` to automatically create
and delete Boxi desktop files for Toolbox containers.

This is something akin to systemd's generator concept, applied to desktop
files.

It's not exactly clear yet how this would (or should) work in context of
distributing Boxi as a Flatpak.  It might be better to do this from
Toolbox itself.