Clean up handling of session close
When the session (from the agent) closes, make sure we close the socket
and drop our "listener" reference on the window.

We also don't need to handle the "eof" signal from Vte, nor do we need
to destroy the window on exit.  The only place we should destroy the
window from is when the session closes.